Made refactor code
