+++ .gitignore
@@ -5,3 +5,4 @@
 /yarn.lock
 /out
 /build
+.env
