+++ src/app/user-posts/page.tsx
@@ -1,0 +1,7 @@
+import UserPostComponent from "../components/UserPostComponent";
+
+const UserPosts = () => {
+  return <UserPostComponent />;
+};
+
+export default UserPosts;
